// packages/contracts/generated/OntologyVersion.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Id } from "./Id";
import type { LabelDef } from "./LabelDef";

export type OntologyVersion = { ontology_version_id: Id, project_id: Id, version_no: number, labels: Array<LabelDef>, guidelines_markdown: string, allow_out_of_bounds: false, };